Fix user API
