+++ views/options.ejs
@@ -14,10 +14,6 @@
     </script>
 </head>
 <body class="full">
-<div class='tooltip' role='tooltip'>
-    <div class='tooltip-arrow'></div>
-    <div class='tooltip-inner'>Hello my name is Daan</div>
-</div>
 <div id='wrap' class='full'>
     <% include partials/general_navigationbar.ejs %>
     <% include partials/alertMessage.ejs %>
