+++ views/partials/manageAdmins.ejs
@@ -10,7 +10,7 @@
         <div class='tab-content'>
             <div role='tabpanel' class='tab-pane fade in active' id='home'>
                 <h3>Add an administrator</h3>
-                <form method='post' enctype='application/json'>
+                <form method='post' enctype='application/json' action='/admin'>
                     <div class='modal-body modal-add-body'>
 
                         <div class='form-group'>
